Move webpack config to config directory
